ISSUES-4006 split protocol text packet
