+++ src/Core/MySQL/IMySQLReadPacket.cpp
@@ -37,6 +37,38 @@ void LimitedReadPacket::readPayloadWithUnpacked(ReadBuffer & in)
     IMySQLReadPacket::readPayloadWithUnpacked(limited);
 }
 
+uint64_t readLengthEncodedNumber(ReadBuffer & buffer)
+{
+    char c{};
+    uint64_t buf = 0;
+    buffer.readStrict(c);
+    auto cc = static_cast<uint8_t>(c);
+    if (cc < 0xfc)
+    {
+        return cc;
+    }
+    else if (cc < 0xfd)
+    {
+        buffer.readStrict(reinterpret_cast<char *>(&buf), 2);
+    }
+    else if (cc < 0xfe)
+    {
+        buffer.readStrict(reinterpret_cast<char *>(&buf), 3);
+    }
+    else
+    {
+        buffer.readStrict(reinterpret_cast<char *>(&buf), 8);
+    }
+    return buf;
+}
+
+void readLengthEncodedString(String & s, ReadBuffer & buffer)
+{
+    uint64_t len = readLengthEncodedNumber(buffer);
+    s.resize(len);
+    buffer.readStrict(reinterpret_cast<char *>(s.data()), len);
+}
+
 }
 
 }
