+++ src/Core/MySQL/MySQLPackets.cpp
@@ -235,75 +235,6 @@ void ResponsePacket::readPayloadImpl(ReadBuffer & payload)
     }
 }
 
-ColumnDefinitionPacket::ColumnDefinitionPacket()
-    : character_set(0x00), column_length(0), column_type(MYSQL_TYPE_DECIMAL), flags(0x00)
-{
-}
-
-ColumnDefinitionPacket::ColumnDefinitionPacket(
-    String schema_, String table_, String org_table_, String name_, String org_name_, uint16_t character_set_, uint32_t column_length_,
-    ColumnType column_type_, uint16_t flags_, uint8_t decimals_)
-    : schema(std::move(schema_)), table(std::move(table_)), org_table(std::move(org_table_)), name(std::move(name_)),
-      org_name(std::move(org_name_)), character_set(character_set_), column_length(column_length_), column_type(column_type_),
-      flags(flags_), decimals(decimals_)
-{
-}
-
-ColumnDefinitionPacket::ColumnDefinitionPacket(
-    String name_, uint16_t character_set_, uint32_t column_length_, ColumnType column_type_, uint16_t flags_, uint8_t decimals_)
-    : ColumnDefinitionPacket("", "", "", std::move(name_), "", character_set_, column_length_, column_type_, flags_, decimals_)
-{
-}
-
-size_t ColumnDefinitionPacket::getPayloadSize() const
-{
-    return 13 + getLengthEncodedStringSize("def") + getLengthEncodedStringSize(schema) + getLengthEncodedStringSize(table) + getLengthEncodedStringSize(org_table) + \
-            getLengthEncodedStringSize(name) + getLengthEncodedStringSize(org_name) + getLengthEncodedNumberSize(next_length);
-}
-
-void ColumnDefinitionPacket::readPayloadImpl(ReadBuffer & payload)
-{
-    String def;
-    readLengthEncodedString(def, payload);
-    assert(def == "def");
-    readLengthEncodedString(schema, payload);
-    readLengthEncodedString(table, payload);
-    readLengthEncodedString(org_table, payload);
-    readLengthEncodedString(name, payload);
-    readLengthEncodedString(org_name, payload);
-    next_length = readLengthEncodedNumber(payload);
-    payload.readStrict(reinterpret_cast<char *>(&character_set), 2);
-    payload.readStrict(reinterpret_cast<char *>(&column_length), 4);
-    payload.readStrict(reinterpret_cast<char *>(&column_type), 1);
-    payload.readStrict(reinterpret_cast<char *>(&flags), 2);
-    payload.readStrict(reinterpret_cast<char *>(&decimals), 2);
-    payload.ignore(2);
-}
-
-void ColumnDefinitionPacket::writePayloadImpl(WriteBuffer & buffer) const
-{
-    writeLengthEncodedString(std::string("def"), buffer); /// always "def"
-    writeLengthEncodedString(schema, buffer);
-    writeLengthEncodedString(table, buffer);
-    writeLengthEncodedString(org_table, buffer);
-    writeLengthEncodedString(name, buffer);
-    writeLengthEncodedString(org_name, buffer);
-    writeLengthEncodedNumber(next_length, buffer);
-    buffer.write(reinterpret_cast<const char *>(&character_set), 2);
-    buffer.write(reinterpret_cast<const char *>(&column_length), 4);
-    buffer.write(reinterpret_cast<const char *>(&column_type), 1);
-    buffer.write(reinterpret_cast<const char *>(&flags), 2);
-    buffer.write(reinterpret_cast<const char *>(&decimals), 2);
-    writeChar(0x0, 2, buffer);
-}
-
-void ComFieldList::readPayloadImpl(ReadBuffer & payload)
-{
-    // Command byte has been already read from payload.
-    readNullTerminated(table, payload);
-    readStringUntilEOF(field_wildcard, payload);
-}
-
 LengthEncodedNumber::LengthEncodedNumber(uint64_t value_) : value(value_)
 {
 }
