+++ src/Core/MySQL/PacketsConnection.h
@@ -34,7 +34,9 @@ protected:
 public:
     Handshake();
 
-    Handshake(uint32_t capability_flags_, uint32_t connection_id_, String server_version_, String auth_plugin_name_, String auth_plugin_data_);
+    Handshake(
+        uint32_t capability_flags_, uint32_t connection_id_,
+        String server_version_, String auth_plugin_name_, String auth_plugin_data_, uint8_t charset_);
 };
 
 class HandshakeResponse : public IMySQLWritePacket, public IMySQLReadPacket
