+++ src/Core/MySQL/PacketsProtocolText.cpp
@@ -1,5 +1,7 @@
 #include <Core/MySQL/PacketsProtocolText.h>
-#include <Core/MySQLProtocol.h>
+#include <IO/WriteBufferFromString.h>
+#include <IO/ReadHelpers.h>
+#include <IO/WriteHelpers.h>
 
 namespace DB
 {
@@ -44,6 +46,143 @@ void ResultSetRow::writePayloadImpl(WriteBuffer & buffer) const
         else
             writeLengthEncodedString(serialized[i], buffer);
     }
+}
+
+void ComFieldList::readPayloadImpl(ReadBuffer & payload)
+{
+    // Command byte has been already read from payload.
+    readNullTerminated(table, payload);
+    readStringUntilEOF(field_wildcard, payload);
+}
+
+ColumnDefinition::ColumnDefinition()
+    : character_set(0x00), column_length(0), column_type(MYSQL_TYPE_DECIMAL), flags(0x00)
+{
+}
+
+ColumnDefinition::ColumnDefinition(
+    String schema_, String table_, String org_table_, String name_, String org_name_, uint16_t character_set_, uint32_t column_length_,
+    ColumnType column_type_, uint16_t flags_, uint8_t decimals_)
+    : schema(std::move(schema_)), table(std::move(table_)), org_table(std::move(org_table_)), name(std::move(name_)),
+      org_name(std::move(org_name_)), character_set(character_set_), column_length(column_length_), column_type(column_type_),
+      flags(flags_), decimals(decimals_)
+{
+}
+
+ColumnDefinition::ColumnDefinition(
+    String name_, uint16_t character_set_, uint32_t column_length_, ColumnType column_type_, uint16_t flags_, uint8_t decimals_)
+    : ColumnDefinition("", "", "", std::move(name_), "", character_set_, column_length_, column_type_, flags_, decimals_)
+{
+}
+
+size_t ColumnDefinition::getPayloadSize() const
+{
+    return 13 + getLengthEncodedStringSize("def") + getLengthEncodedStringSize(schema) + getLengthEncodedStringSize(table) + getLengthEncodedStringSize(org_table) + \
+            getLengthEncodedStringSize(name) + getLengthEncodedStringSize(org_name) + getLengthEncodedNumberSize(next_length);
+}
+
+void ColumnDefinition::readPayloadImpl(ReadBuffer & payload)
+{
+    String def;
+    readLengthEncodedString(def, payload);
+    assert(def == "def");
+    readLengthEncodedString(schema, payload);
+    readLengthEncodedString(table, payload);
+    readLengthEncodedString(org_table, payload);
+    readLengthEncodedString(name, payload);
+    readLengthEncodedString(org_name, payload);
+    next_length = readLengthEncodedNumber(payload);
+    payload.readStrict(reinterpret_cast<char *>(&character_set), 2);
+    payload.readStrict(reinterpret_cast<char *>(&column_length), 4);
+    payload.readStrict(reinterpret_cast<char *>(&column_type), 1);
+    payload.readStrict(reinterpret_cast<char *>(&flags), 2);
+    payload.readStrict(reinterpret_cast<char *>(&decimals), 2);
+    payload.ignore(2);
+}
+
+void ColumnDefinition::writePayloadImpl(WriteBuffer & buffer) const
+{
+    writeLengthEncodedString(std::string("def"), buffer); /// always "def"
+    writeLengthEncodedString(schema, buffer);
+    writeLengthEncodedString(table, buffer);
+    writeLengthEncodedString(org_table, buffer);
+    writeLengthEncodedString(name, buffer);
+    writeLengthEncodedString(org_name, buffer);
+    writeLengthEncodedNumber(next_length, buffer);
+    buffer.write(reinterpret_cast<const char *>(&character_set), 2);
+    buffer.write(reinterpret_cast<const char *>(&column_length), 4);
+    buffer.write(reinterpret_cast<const char *>(&column_type), 1);
+    buffer.write(reinterpret_cast<const char *>(&flags), 2);
+    buffer.write(reinterpret_cast<const char *>(&decimals), 2);
+    writeChar(0x0, 2, buffer);
+}
+
+ColumnDefinition getColumnDefinition(const String & column_name, const TypeIndex type_index)
+{
+    ColumnType column_type;
+    CharacterSet charset = CharacterSet::binary;
+    int flags = 0;
+    switch (type_index)
+    {
+        case TypeIndex::UInt8:
+            column_type = ColumnType::MYSQL_TYPE_TINY;
+            flags = ColumnDefinitionFlags::BINARY_FLAG | ColumnDefinitionFlags::UNSIGNED_FLAG;
+            break;
+        case TypeIndex::UInt16:
+            column_type = ColumnType::MYSQL_TYPE_SHORT;
+            flags = ColumnDefinitionFlags::BINARY_FLAG | ColumnDefinitionFlags::UNSIGNED_FLAG;
+            break;
+        case TypeIndex::UInt32:
+            column_type = ColumnType::MYSQL_TYPE_LONG;
+            flags = ColumnDefinitionFlags::BINARY_FLAG | ColumnDefinitionFlags::UNSIGNED_FLAG;
+            break;
+        case TypeIndex::UInt64:
+            column_type = ColumnType::MYSQL_TYPE_LONGLONG;
+            flags = ColumnDefinitionFlags::BINARY_FLAG | ColumnDefinitionFlags::UNSIGNED_FLAG;
+            break;
+        case TypeIndex::Int8:
+            column_type = ColumnType::MYSQL_TYPE_TINY;
+            flags = ColumnDefinitionFlags::BINARY_FLAG;
+            break;
+        case TypeIndex::Int16:
+            column_type = ColumnType::MYSQL_TYPE_SHORT;
+            flags = ColumnDefinitionFlags::BINARY_FLAG;
+            break;
+        case TypeIndex::Int32:
+            column_type = ColumnType::MYSQL_TYPE_LONG;
+            flags = ColumnDefinitionFlags::BINARY_FLAG;
+            break;
+        case TypeIndex::Int64:
+            column_type = ColumnType::MYSQL_TYPE_LONGLONG;
+            flags = ColumnDefinitionFlags::BINARY_FLAG;
+            break;
+        case TypeIndex::Float32:
+            column_type = ColumnType::MYSQL_TYPE_FLOAT;
+            flags = ColumnDefinitionFlags::BINARY_FLAG;
+            break;
+        case TypeIndex::Float64:
+            column_type = ColumnType::MYSQL_TYPE_DOUBLE;
+            flags = ColumnDefinitionFlags::BINARY_FLAG;
+            break;
+        case TypeIndex::Date:
+            column_type = ColumnType::MYSQL_TYPE_DATE;
+            flags = ColumnDefinitionFlags::BINARY_FLAG;
+            break;
+        case TypeIndex::DateTime:
+            column_type = ColumnType::MYSQL_TYPE_DATETIME;
+            flags = ColumnDefinitionFlags::BINARY_FLAG;
+            break;
+        case TypeIndex::String:
+        case TypeIndex::FixedString:
+            column_type = ColumnType::MYSQL_TYPE_STRING;
+            charset = CharacterSet::utf8_general_ci;
+            break;
+        default:
+            column_type = ColumnType::MYSQL_TYPE_STRING;
+            charset = CharacterSet::utf8_general_ci;
+            break;
+    }
+    return ColumnDefinition(column_name, charset, 0, column_type, flags, 0);
 }
 
 }
